refactor: Commitsコンポーネントを分割した
